pyproject.toml
[project]
name = "diffusion-posterior-sampling"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
  "torch==2.11.0+cu130",
  "torchvision==0.26.0+cu130",
  "torchaudio==2.11.0+cu130",
  "lpips==0.1.4",
  "matplotlib==3.10.9",
  "numpy==2.4.4",
  "Pillow==12.2.0",
  "PyYAML==6.0.3",
  "scikit-image==0.26.0",
  "scipy==1.17.1",
  "tqdm==4.67.3",
  "wandb>=0.26.1",
]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu130" }
torchvision = { index = "pytorch-cu130" }
torchaudio = { index = "pytorch-cu130" }

